+++ bindings/python/test/interpolators/test_barycentric_rational.py
@@ -7,7 +7,7 @@
 
 ################################################################################################################################################################
 
-from ostk.mathematics.curve_fitting.interpolation import BarycentricRational
+from ostk.mathematics.curve_fitting.interpolator import BarycentricRational
 
 ################################################################################################################################################################
