+++ bindings/python/test/interpolators/test_linear.py
@@ -7,7 +7,7 @@
 
 ################################################################################################################################################################
 
-from ostk.mathematics.curve_fitting.interpolation import Linear
+from ostk.mathematics.curve_fitting.interpolator import Linear
 
 ################################################################################################################################################################
