+++ include/OpenSpaceToolkit/Mathematics/CurveFitting/Interpolator/BarycentricRational.hpp
@@ -1,22 +1,21 @@
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 /// @project        Open Space Toolkit ▸ Mathematics
-/// @file           OpenSpaceToolkit/Mathematics/Interpolation/Linear.hpp
+/// @file           OpenSpaceToolkit/Mathematics/CurveFitting/Interpolator/BarycentricRational.hpp
 /// @author         Vishwa Shah <vishwa@loftorbital.com>
 /// @license        Apache License 2.0
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-#ifndef __OpenSpaceToolkit_Mathematics_Interpolator_Linear__
-#define __OpenSpaceToolkit_Mathematics_Interpolator_Linear__
+#ifndef __OpenSpaceToolkit_Mathematics_Interpolator_BarycentricRational__
+#define __OpenSpaceToolkit_Mathematics_Interpolator_BarycentricRational__
 
 #include <OpenSpaceToolkit/Core/Types/Real.hpp>
-#include <OpenSpaceToolkit/Core/Types/Size.hpp>
-#include <OpenSpaceToolkit/Core/Types/Index.hpp>
-#include <OpenSpaceToolkit/Core/Containers/Pair.hpp>
 
 #include <OpenSpaceToolkit/Mathematics/Objects/Vector.hpp>
 #include <OpenSpaceToolkit/Mathematics/CurveFitting/Interpolator.hpp>
+
+#include <boost/math/interpolators/barycentric_rational.hpp>
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -33,22 +32,22 @@ namespace interp
 
 using ostk::core::types::Real ;
 using ostk::core::types::Size ;
-using ostk::core::types::Index ;
-using ostk::core::ctnr::Pair ;
 
 using ostk::math::obj::VectorXd ;
 using ostk::math::curvefitting::interp::Interpolator ;
 
+using boost::math::interpolators::barycentric_rational ;
+
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-/// @brief                      Linear
+/// @brief                      BarycentricRational
 ///
-///                             In mathematics, linear interpolation is a method of curve fitting using linear polynomials to construct new data
-///                             points within the range of a discrete set of known data points.
+///                             Barycentric rational interpolator is a high-accuracy interpolator method for non-uniformly spaced samples.
+///                             It requires 𝑶(N) time for construction, and 𝑶(N) time for each evaluation.
 ///
-/// @ref                        https://en.wikipedia.org/wiki/Linear_interpolation#:~:text=In%20mathematics%2C%20linear%20interpolation%20is,set%20of%20known%20data%20points.
+/// @ref                        https://www.boost.org/doc/libs/1_81_0/libs/math/doc/html/math_toolkit/barycentric.html
 
-class Linear: public Interpolator
+class BarycentricRational: public Interpolator
 {
 
     public:
@@ -56,7 +55,7 @@ class Linear: public Interpolator
         /// @brief              Constructor
         ///
         /// @code
-        ///                     Linear linear(x, y) ;
+        ///                     BarycentricRational barycentricRational(x, y) ;
         /// @endcode
         ///
         /// @param              [in] anXVector A vector of x values
@@ -65,30 +64,30 @@ class Linear: public Interpolator
         /// @warning            The x values must be sorted in ascending order
         /// @warning            The x values must be equally spaced
 
-                                Linear                                      (   const   VectorXd&                   anXVector,
+                                BarycentricRational                         (   const   VectorXd&                   anXVector,
                                                                                 const   VectorXd&                   aYVector                                    ) ;
 
-        /// @brief              Clone linear
+        /// @brief              Clone barycentric rational
         ///
-        /// @return             Pointer to cloned linear interpolator
+        /// @return             Pointer to cloned barycentric rational
 
-        virtual Linear*         clone                                ( ) const ;
+        BarycentricRational*    clone                                       ( ) const ;
 
-        /// @brief              Evaluate the linear interpolator
+        /// @brief              Evaluate the spline
         ///
         /// @code
-        ///                     VectorXd values = linear.evaluate({1.0, 5.0, 6.0}) ;
+        ///                     VectorXd values = barycentricRational.evaluate({1.0, 5.0, 6.0}) ;
         /// @endcode
         ///
         /// @param              [in] aQueryVector A vector of x values
         /// @return             Vector of y values
 
         VectorXd                evaluate                                    (   const   VectorXd&                   aQueryVector                                ) const ;
-
-        /// @brief              Evaluate the linear interpolator
+    
+        /// @brief              Evaluate the spline
         ///
         /// @code
-        ///                     double values = linear.evaluate(5.0) ;
+        ///                     double values = barycentricRational.evaluate(5.0) ;
         /// @endcode
         ///
         /// @param              [in] aQueryValue An x value
@@ -98,10 +97,7 @@ class Linear: public Interpolator
 
     private:
 
-        VectorXd                x_ ;
-        VectorXd                y_ ;
-
-        Pair<Index, Index>      findIndexRange                              (   const   double&                     aQueryValue                                 ) const ;
+        barycentric_rational<double> interpolator_ ;
 
 } ;
